Last working version. Better expression for the angular velocity
The angular velocity was improved employing the correct definition of the time derivatives of Euler angles. Still to resolve an issue with the initial conditions in terms of Euler angles, the expression obtained from director cosines are not unique. the consequence of this is that molecule and body fixed frame are not aligned. Dynamically there are not problems, only graphically.